// frontend/src/sections/stock-movements/list/insertStockMovement.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../../api/graphql";

export type InsertStockMovementVariables = {
  storeId: string;
  input: {
    id: string;
    comment?: string | null;
  };
};

export type InsertStockMovementResult = {
  insertStockRelocation: ({
  id: string;
});
};

export const InsertStockMovement = {
  query: "mutation InsertStockMovement($storeId: String!, $input: InsertStockRelocationInput!) {\n  insertStockRelocation(storeId: $storeId, input: $input) {\n    ... on StockRelocationNode {\n      id\n    }\n  }\n}",
} as TypedDocument<InsertStockMovementResult, InsertStockMovementVariables>;